chore: #49 코드리뷰 반영
